Fix libfsimage build on NetBSD.
Fixes a number of these errors:
 cc1: warnings being treated as errors
 fsys_fat.c: In function 'fat_dir':
 fsys_fat.c:304: warning: array subscript has type 'char'


These ugly casts are needed according to the ISO C spec.

Acked-by: Keir Fraser <keir@xensource.com>

+++ tools/libfsimage/ext2fs/fsys_ext2fs.c
@@ -594,7 +594,7 @@ ext2fs_dir (fsi_file_t *ffi, char *dirname)
 
 	  /* Find out how long our remaining name is. */
 	  len = 0;
-	  while (dirname[len] && !isspace (dirname[len]))
+	  while (dirname[len] && !isspace ((uint8_t)dirname[len]))
 	    len++;
 
 	  /* Get the symlink size. */
@@ -651,7 +651,7 @@ ext2fs_dir (fsi_file_t *ffi, char *dirname)
 	}
 
       /* if end of filename, INODE points to the file's inode */
-      if (!*dirname || isspace (*dirname))
+      if (!*dirname || isspace ((uint8_t)*dirname))
 	{
 	  if (!S_ISREG (INODE->i_mode))
 	    {
@@ -678,7 +678,7 @@ ext2fs_dir (fsi_file_t *ffi, char *dirname)
 	}
 
       /* skip to next slash or end of filename (space) */
-      for (rest = dirname; (ch = *rest) && !isspace (ch) && ch != '/';
+      for (rest = dirname; (ch = *rest) && !isspace ((uint8_t)ch) && ch != '/';
 	   rest++);
 
       /* look through this directory and find the next filename component */
